feat(cli): add default allow dirty to publish
The publish subcommmand previously did not allow publication if
there were any uncommitted files in the working directory (the
default behaviour of "cargo publish"). Change the default behaviour
to use the "--allow-dirty" flag to "cargo publish" and provide a
"--no-dirty" flag to the publish subcommmand to go back to the
original behaviour.

The default use of "--allow-dirty" is reasonable because the prepare
subcommmand will make changes to the files in the workspace (specifically
the Cargo.toml files). These changes may or may not be committed to the
git repository by the time the publish subcommmand is run.