fix(release): change release workflow to use git plugin
The former method of supporting semantic release attempted to use
a pull request and automatic merging of the that pull request to
commit the changed version numbers to the release branch but this
did not commit at the right time and didn't properly work in an
automated manner. The new method uses the git plugin for semantic
release to accomplish the same goal but at a time that works better
for the subsequent semantic release steps and (hopefully) is
fully automated.

A key change from prior attempts to use the git plugin is the
"persist-credentials: false" argument to the Checkout step. This
appears to be necessary for the git plugin to properly use the
"sbosnick-bot" personal access token from secrets.SEMRELGH.

This change also refactors the push trigger out of the ci workflow
into its own release workflow. The ci workflow is now exclusivly for
pull requests.